Change how database is built
