+++ src/Notifications.API/Database/AppDbContext.cs
@@ -3,15 +3,6 @@ using Notifications.Models;
 
 namespace Notifications.Database;
 
-internal static class DatabaseExtensions
-{
-    public static void ConfigureDatabase(this WebApplicationBuilder builder)
-    {
-        builder.Services.AddDbContext<AppDbContext>(options =>
-            options.UseSqlite(builder.Configuration.GetConnectionString("DefaultConnection")));
-    }
-}
-
 internal class AppDbContext(DbContextOptions<AppDbContext> options) : DbContext(options)
 {
     public DbSet<Notification> Notifications { get; set; } = null!;
