+++ src/Notifications.API/Database/TestData.cs
@@ -1,23 +1,13 @@
-﻿using Notifications.Models;
+﻿using Microsoft.EntityFrameworkCore;
+using Notifications.Models;
 
 namespace Notifications.Database;
 
 internal static class TestData
 {
-    public static async Task BuildDatabaseInDevAsync(this WebApplication app)
-    {
-        if (!app.Environment.IsDevelopment()) return;
-
-        using var scope = app.Services.CreateScope();
-        var context = scope.ServiceProvider.GetRequiredService<AppDbContext>();
-        await context.Database.EnsureDeletedAsync();
-        await context.Database.EnsureCreatedAsync();
-        await SeedDataAsync(context);
-    }
-
     public static async Task SeedDataAsync(AppDbContext context)
     {
-        if (context.Notifications.Any()) return;
+        if (await context.Notifications.AnyAsync()) return;
         context.Notifications.AddRange(NotificationSeedItems);
         await context.SaveChangesAsync();
     }
